requirements.txt
torch~=2.0.1+cu117
numpy~=1.24.4
pandas~=1.5.3
typing~=3.10.0.0
wandb~=0.15.5
transformers~=4.29.0
hydra-core==1.2.0
torchvision~=0.15.2+cu117
omegaconf~=2.3.0
sacrebleu~=2.3.1
tqdm~=4.65.0
pillow~=9.3.0
opencv-python~=4.7.0.72
groundingdino~=0.1.0
click~=8.1.3
setuptools~=59.5.0